REFACTOR split handleRequest into two js file and applied changes to react components
